+++ gc_clock.hh
@@ -28,11 +28,10 @@
 #include <chrono>
 #include <optional>
 
-// FIXME: wraps around in 2038
 class gc_clock final {
 public:
     using base = seastar::lowres_system_clock;
-    using rep = int32_t;
+    using rep = int64_t;
     using period = std::ratio<1, 1>; // seconds
     using duration = std::chrono::duration<rep, period>;
     using time_point = std::chrono::time_point<gc_clock, duration>;
